+++ NEWS.md
@@ -1,4 +1,4 @@
-# finnts 0.6.0.9040 (development version)
+# finnts 0.6.0.9041 (development version)
 
 ## Improvements
 
@@ -48,7 +48,8 @@
 -   Fixed weighted mape calculation when target variable has negative values.
 -   Support for latest xgboost 3x version.
 -   Fixed model summary for global models by considering average models too.
-  
+-   Fixed issue when future values of external regressors exist in some series but not all, leading to missing data issues when training a global model.
+
 ## Breaking Changes
 
 - `experiment_name` within `set_run_info()` has been changed to `project_name` to comply with new AI agent capabilities.